# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRM Laravel Api",
        "version": "10.0"
    },
    "paths": {
...
        "/api/equipment": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Get All active equipments.Permission required = equipment.list",
                "description": "This endpoint retrieves information about something.",
                "operationId": "2c44a71ff45eecd201cdcff51b2ba254",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Create a new equipment.Permission required = equipment.store",
                "description": "This endpoint creates a new equipment.",
                "operationId": "1ad80519d8ce98f958195b6e87cd95c6",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The name of the equipment => required",
                                        "type": "string",
                                        "example": "gun"
                                    },
                                    "status": {
                                        "description": "The status of the equipment =>required",
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "equipment created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/equipment/inactive-equipments": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Get All inactive equipments.Permission required = equipment.list",
                "description": "This endpoint retrieves information about something.",
                "operationId": "127ae4a41d42666b9e29532ddaf1dc3c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/equipment/{id}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "GET The equipment.Permission required = equipment.edit",
                "description": "This endpoint Gives a specific  equipment.",
                "operationId": "faf3b9e534573c63c8cfe9244ad05234",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the equipment ",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Delete The equipment.Permission required = equipment.delete",
                "description": "This endpoint delete equipment.",
                "operationId": "20db7db0d4d1ad4195f1a893d4d6036a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the equipment to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Update the equipment.Permission required = equipment.update",
                "description": "This endpoint updates a equipment.",
                "operationId": "dca260a92651ce3f499e9315503f2845",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the equipment to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The name of the equipment => required",
                                        "type": "string",
                                        "example": "gun"
                                    },
                                    "status": {
                                        "description": "The status of the equipment =>required",
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "equipment updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}